requirements.txt
anyio==4.4.0
fastapi==0.112.0
fastapi-cli==0.0.5
openai==1.38.0
python-dotenv==1.0.1
uvicorn==0.30.5
uvloop==0.19.0
websockets==12.0
PyGithub